service refactor and fixes: outside world resources are now services, some coupling left with terminal, doesn't worth the fight, and now I know why not everything is PnP, i'm boiler-plate overloaded.

+++ render/renderers/heat_meter.go
@@ -29,7 +29,7 @@ func NewHeatMeterRenderer(ctx *engine.GameContext) *HeatMeterRenderer {
 	// Access RenderConfig for display mode
 	cfg := engine.MustGetResource[*engine.RenderConfig](ctx.World.Resources)
 
-	if cfg.ColorMode == uint8(terminal.ColorMode256) {
+	if cfg.ColorMode == terminal.ColorMode256 {
 		h.renderCell = h.cell256
 	} else {
 		h.renderCell = h.cellTrueColor
